backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
dj-database-url==2.1.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# Async & WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Billing
stripe==7.8.0

# HTTP & WebSocket clients
requests==2.31.0
websockets==12.0
aiohttp==3.9.1

# AI & ML
numpy==1.24.4
scikit-learn==1.3.2

# Utilities
python-decouple==3.8
Pillow==10.1.0
python-dateutil==2.8.2

# Development
black==23.11.0
ruff==0.1.6
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Logging
loguru==0.7.2

# Security
cryptography>=41.0.0
django-cryptography==1.1
